Add basic structure to index.html
Add in my planned structure to `index.html'

I plan for the page to have a title for the climbing centre
and to have the site navigation underneath.

I will then have a short description of the site and centre,
followed by a ficticious location for the centre.

The page will then finish with a footer containing
social media links as well as a copyright notice.