Add section `Weekly events' to plan for description section
This adds a plan to the structure of the welcome page to include a
section which will list the weekly events taking place at the centre.

I'm thinking of social events, including ones where
people who might be new to climbing or
who don't come to the centre with a climbing partner
can meet and be paired.
Another kind of social event might be more competitive climbing,
or organised trips to outdoor, partner sites.