Set the iframe's border in CSS instead of HTML
* Remove the attribute setting the frameborder
This is no longer a recommended way to set the border of an element.
The modern approach is to use CSS.

As well as the use of CSS being recommended, the frameborder attribute
is now deprecated for iframes.

* Set the border of the iframe in CSS
This is the recommended way to set the border of an element.