Wrap info for 'about us' section in a div
I want to use CSS grid on the 'about us' seciton but I don't want to
include the heading in the grid.

I could have the heading in the grid and set it to be on the first
row and to take up all the columns, but (at least for now) I've
seperated the heading from the main contents of the section.