Remove preceding whitespace from 'who are we' paragraph
This might be considered to be odd formating for HTML,
but the reason I do it is to help keep the code on the left side
of the screen.
I believe this helps with readability, as I'm more
comfortable reading multiple lines of text which are justified to
the left than reading long lines of text.

With this, the preceding whitespace makes no difference to the
rendered HTML document, as whitespace in a paragraph element
is ignored.