Move 'weekly events' heading out of article
I ran my HTML through the W3C validator.
It warned that I had an section which did not have a heading element.

I don't think that the 'weekly events' section fits the definition
of an article, so I'm happy remove the article and just have it inside
a section element.