Add an image to break apart the 'about us' and 'why you should climb' sections
I've added an image to seperate the 'about us' and
'why you should climb' sections.
I've styled this image so that for viewport widths smaller than
768px, the images height will be about a third of the viewports width.
For viewports that are wider, it is fixed to 320px in height.

It's also set to cover the div area it's the background to.
This gives the effect of it zooming in on the image when the viewport
is wider than 767px, otherwise the aspect ratio remains the same.